#368: Refactored further to remove internal models from models package.
