+++ internal/export/encryption_test.go
@@ -4,7 +4,7 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-package models
+package distro
 
 import (
 	"crypto/aes"
