+++ internal/export/registration_test.go
