--- pkg/models/filter.go
@@ -1,13 +1,0 @@
-//
-// Copyright (c) 2017 Mainflux
-//
-// SPDX-License-Identifier: Apache-2.0
-//
-
-package models
-
-// Filter - Specifies the client filters on reading data
-type Filter struct {
-	DeviceIDs          []string `bson:"deviceIdentifiers,omitempty" json:"deviceIdentifiers,omitempty"`
-	ValueDescriptorIDs []string `bson:"valueDescriptorIdentifiers,omitempty" json:"valueDescriptorIdentifiers,omitempty"`
-}
